+++ ParserAndroidPrototype/src/game/EntityMouvementBehavior/Harmonic.java
@@ -17,10 +17,10 @@ public class Harmonic extends EntityMouvementBehavior
 	@Override
 	public void update(Entity e1, double heartBeat, double dt) 
 	{
-		e1.setVelX( Math.cos(Math.toRadians( getAngle())) * startDir.x +
-				( startDir.x * dt  * Acceleration.ROLL.getdBound() ) );
-		e1.setVelY( Math.sin(Math.toRadians(getAngle())) * startDir.y + 
-				( startDir.y * dt * Acceleration.ROLL.getdBound() ) );
+		e1.setVelX( (float) (Math.cos(Math.toRadians( getAngle())) * startDir.x +
+				( startDir.x * dt  * Acceleration.ROLL.getdBound() )) );
+		e1.setVelY( (float) (Math.sin(Math.toRadians(getAngle())) * startDir.y + 
+				( startDir.y * dt * Acceleration.ROLL.getdBound() )) );
 		
 		e1.setPosX((int)(e1.getPosX() + e1.getVelX()));
